+++ src/main/java/com/deal4u/fourplease/domain/member/entity/Member.java
@@ -1,6 +1,7 @@
 package com.deal4u.fourplease.domain.member.entity;
 
-import com.deal4u.fourplease.domain.BaseDateEntity;
+
+import com.deal4u.fourplease.domain.common.BaseDateEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
