+++ src/test/java/com/deal4u/fourplease/domain/auction/controller/SaleControllerTests.java
@@ -13,20 +13,21 @@ import com.deal4u.fourplease.domain.auction.service.AuctionService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-@WebMvcTest(SaleController.class)
+@SpringBootTest
+@AutoConfigureMockMvc(addFilters = false)
 class SaleControllerTests {
-
-    @Autowired
-    private MockMvc mockMvc;
 
     @MockitoBean
     AuctionService auctionService;
+    @Autowired
+    private MockMvc mockMvc;
 
     @Test
     @DisplayName("GET /api/v1/sales/{sellerId}이 성공하면 id에 해당하는 판매자의 정보와 200을 반환한다")
@@ -40,10 +41,10 @@ class SaleControllerTests {
         when(auctionService.findSalesBySellerId(sellerId, pageable)).thenReturn(resp);
 
         mockMvc.perform(
-                get("/api/v1/sales/{sellerId}", sellerId)
-                        .param("page", "0")
-                        .param("size", "20")
-        ).andExpect(status().isOk())
+                        get("/api/v1/sales/{sellerId}", sellerId)
+                                .param("page", "0")
+                                .param("size", "20")
+                ).andExpect(status().isOk())
                 .andExpect(jsonPath("$.content[0].name")
                         .value(resp.getContent().get(0).name()))
                 .andExpect(jsonPath("$.content[1].name")
@@ -56,7 +57,6 @@ class SaleControllerTests {
                 .andExpect(jsonPath("$.size").value(resp.getSize()))
                 .andDo(print());
     }
-
 
 
 }
